requirements.txt
SQLAlchemy~=1.4.37
aiogram~=2.20
python-dotenv~=0.20.0
psycopg2==2.9.3
